# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: LearnFlow API
  version: 1.0.0
  description: AI-powered learning platform API
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/profile/context:
    get:
      summary: Get Student Context Object
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Student context
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/profile/goals:
    post:
      summary: Save learning goals
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Saved
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/profile/onboarding/complete:
    post:
      summary: Mark onboarding complete
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Updated
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/profile/data-summary:
    get:
      summary: Get profile data summary
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Summary
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                properties:
                  coursesCount: { type: integer }
                  pipelinesCount: { type: integer }
                  notificationsCount: { type: integer }
                  lastCourseCreatedAt: { type: string, nullable: true }
                  lastPipelineCreatedAt: { type: string, nullable: true }
                  lastNotificationCreatedAt: { type: string, nullable: true }
        '401':
          description: Unauthorized
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/notifications:
    get:
      summary: List notifications
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Notifications
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/notifications/read:
    post:
      summary: Mark notification read
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
      responses:
        '200':
          description: OK
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/notifications/read-all:
    post:
      summary: Mark all notifications as read
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Empty object
      responses:
        '200':
          description: OK
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  updated: { type: integer }
        '400':
          description: Validation error
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/notifications/generate:
    post:
      summary: Generate notifications (MVP)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                topic: { type: string }
      responses:
        '201':
          description: Created
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Conflict
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    XRequestId:
      description: Request identifier for tracing
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        requestId:
          type: string
        status:
          type: integer
...